Allow for compilation outside the root
This allows the default twig to successfully compile with the
following command:

phr up src public_html

Where src is the source directory and public_html is the target.

Which in my not so humble opinion is a better practice than a
hidden subdirectory in the webroot for the source.